Rename gitRepositories query
